+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node app.js",
+    "start": "node server.js",
     "migrate": "npx sequelize db:migrate",
     "seed": "npx sequelize db:seed:all",
     "reset": "npx sequelize db:migrate:undo:all && npx sequelize db:migrate",
