+++ routes/api/models/user.js
@@ -1,7 +1,7 @@
 const path = require("path");
-const sessionHelper = require("../../helpers/session.js")
+const sessionHelper = require(rootPath + "/session/sessionController")
 
-const db = require("../../models")
+const db = require(rootPath + "/models")
 
 module.exports = function (app) {
