+++ routes/html/htmlRoutes.js
@@ -1,0 +1,7 @@
+module.exports = function (app) {
+
+    app.get("/", function (req, res) {
+        res.render("app");
+    });
+
+};
