--- views/template.pug
@@ -1,22 +1,0 @@
-    
-doctype html
-head
-  meta(charset='UTF-8')
-  meta(name='viewport' content='width=device-width, initial-scale=1.0')
-  meta(http-equiv='X-UA-Compatible' content='ie=edge')
-  title= title
-  link(rel='stylesheet' href='/css/reset.css')
-  link(rel='stylesheet' href='https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css')
-  link(rel='stylesheet' href='/css/styles.css')
-  link(rel='stylesheet' href='/css/typography.css')
-// Navbar
-// Content
-#wrap.h-100
-  |!{ content }
-link(rel='stylesheet' href='https://use.fontawesome.com/releases/v5.8.2/css/all.css')
-script(src='https://code.jquery.com/jquery-3.3.1.min.js')
-script(src='https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js')
-script(src='https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js')
-script(src='/js/app.js')
-script
-  |!{ scripts }
